moved all components into parent dir with same name, dont mount App to dom in tests
