+++ src/imports/modules/auth/client/components/EnrollAccountForm/EnrollAccountForm.test.js
@@ -1,6 +1,6 @@
 import { mount } from 'avoriaz';
 import EnrollAccountForm from './EnrollAccountForm.vue';
-import { setupVue } from './../../../../startup/client/client-index';
+import { setupVue } from '/src/imports/startup/client/client-index';
 import { getActions } from '/tests/unit-test-setup/vuex-alt-test-util';
 import nextTick from 'timeout-as-promise';
