+++ src/imports/modules/auth/client/components/PasswordResetEmailForm/PasswordResetEmailForm.vue
@@ -22,8 +22,8 @@
 
 <script>
 import { mapState, mapActions } from 'vuex-alt';
-import { maskEmailAddress } from './../lib/user-data-masking';
-import AuthError from './AuthError';
+// import { maskEmailAddress } from './../../lib/user-data-masking';
+import AuthError from './../AuthError/AuthError.vue';
 
 export default {
   name: 'PasswordResetEmailForm',
