+++ src/imports/modules/auth/client/components/SignOutBtn/SignOutBtn.vue
