+++ src/imports/modules/auth/client/components/SignUpBtn/SignUpBtn.vue
