+++ src/imports/modules/layout/client/components/AppNavigation/AppNavigation.vue
@@ -46,7 +46,7 @@ a {
 </template>
 
 <script>
-import NavBarAuth from '/src/imports/modules/auth/client/components/NavBarAuth.vue';
+import NavBarAuth from '/src/imports/modules/auth/client/components/NavBarAuth/NavBarAuth.vue';
 
 export default {
   components: {
