+++ src/imports/modules/pages/client/components/PagePrivate/PagePrivate.vue
